Refactor map code
